+++ src/components/DirectorOfDay/directorOfDay.module.less
@@ -1,17 +1,17 @@
 @import "../../../mainStyles/_library.less";
 
-.topDirector_container {
+.directorOfDay_container {
 	color: @color-grey;
 	border: 2px solid @color-red;
 	margin: 0 1%;
-	.topDirector_title_container {
+	.directorOfDay_title_container {
 		background: @color-red;
 		padding: 1% 2.6% 0 2.6%;
 		h2 {
 			color: @color-light;
 		}
 	}
-	.topDirector_content_container {
+	.directorOfDay_content_container {
 		padding: 3% 1% 0 1%;
 		h3 {
 			margin: 0 0 3% 0;
@@ -41,7 +41,7 @@
 			}
 		}
 	}
-	.topDirector_container_content_img_resize {
+	.directorOfDay_container_content_img_resize {
 		margin: 0 auto;
 		width: 50%;
 		overflow: hidden;
