--- src/components/Main/main.module.less
@@ -1,20 +1,0 @@
-@import "../../mainStyles/_library.less";
-	.main_container {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		padding: 0 0 5% 0;
-		color: @color-grey;
-		margin: 0 auto;
-		.title {
-			text-align: center;
-			color: @color-grey;
-			text-decoration: none;
-		}
-		.description {
-			padding: 0 0 4% 0;
-			text-align: center;
-			color: @color-grey;
-			text-decoration: none;
-		}
-	}
