+++ src/components/Video/video.module.less
@@ -31,10 +31,3 @@
     animation: moving 1s;
 }
     
-.ReactModal__Overlay--after-open{
-    opacity: 1;
-}
-    
-.ReactModal__Overlay--before-close{
-    opacity: 0;
-}
